[regexp] Optimizing non nullable plus in Experimental
When compiling a quantifier `body{min, max}`, in the special case where
min > 0, max = infinity and body is not nullable, it can be optimized
by using the last mandatory repetition in the following loop. This
reduces the duplication of the bytecode corresponding to the body.

This work was supervised by Aurèle Barrière, from SYSTEMF at EPFL.