[regexp] Optimizing capture group filtering in Experimental
Instead of dynamically clearing registers during the regexp's
execution, those are filtered after the final match is found. This
changes the bytecode's size from quadratic to linear in the size of the
regexp.

This requires to store, in each thread, the time where each capture
groups and quantifiers were last entered, using a global `clock`. Then,
using a bytecode encoding of the AST containing only capture groups and
quantifiers, we traverse the AST to determine which capture group
should be kept.

This work was supervised by Aurèle Barrière, from SYSTEMF at EPFL.

Bug: v8:14097
Change-Id: Ia6bbbe7b52b4a8a8250ed46f29d9bf139a22f024